[ID-29] refactor(ideaReport): 아이디어 리포트, 아이디어 리포트 메타데이터 생성 리팩토링
