+++ src/main/java/com/github/ideantifyserver/domain/ideareport/dto/request/CreateIdeaReportRequestDto.java
@@ -6,10 +6,30 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+import java.util.UUID;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class CreateIdeaReportRequestDto {
     @NotBlank
     String query;
+
+    List<UUID> keywords;
+
+    @NotBlank
+    String summary;
+
+    @NotBlank
+    String purpose;
+
+    @NotBlank
+    String differentiation;
+
+    @NotBlank
+    String technology;
+
+    @NotBlank
+    String target;
 }
