+++ src/main/java/com/github/ideantifyserver/domain/ideareport/sqs/IdeaReportMetadataResponseListener.java
@@ -1,117 +1,59 @@
 package com.github.ideantifyserver.domain.ideareport.sqs;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.ideantifyserver.domain.ideareport.dto.response.AiIdeaReportResultMessage;
-import com.github.ideantifyserver.domain.ideareport.dto.response.IdeaReportMetadataResponseDto;
-import com.github.ideantifyserver.domain.ideareport.entity.EvaluationScores;
-import com.github.ideantifyserver.domain.ideareport.entity.IdeaReportInput;
-import com.github.ideantifyserver.domain.ideareport.entity.IdeaReportResult;
-import com.github.ideantifyserver.domain.ideareport.entity.IdeaReportResultItem;
-import com.github.ideantifyserver.domain.ideareport.repository.IdeaReportInputRepository;
+import com.github.ideantifyserver.domain.ideareport.dto.response.CreateIdeaReportMetadataResponseDto;
+import com.github.ideantifyserver.domain.ideareport.entity.IdeaReportTask;
+import com.github.ideantifyserver.domain.ideareport.repository.IdeaReportTaskRepository;
 import io.awspring.cloud.sqs.annotation.SqsListener;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.Message;
+import org.springframework.messaging.MessageHeaders;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Comparator;
 import java.util.UUID;
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
 public class IdeaReportMetadataResponseListener {
-
-    private final IdeaReportInputRepository inputRepository;
+    private final IdeaReportTaskRepository ideaReportTaskRepository;
     private final SimpMessagingTemplate messagingTemplate;
-    private final ObjectMapper om;
+    private final ObjectMapper objectMapper;
 
-    private static final String HDR_TYPE = "messageType";
-    private static final String HDR_INPUT_ID = "inputId";
-    private static final String TYPE_RESULT = "IDEA_REPORT_RESULT";
+    private static final String HDR_MESSAGE_TYPE = "messageType";
+    private static final String MESSAGE_TYPE     = "IDEA_REPORT";
 
-    @Transactional
     @SqsListener("${app.sqs.idea-report-metadata.response-queue}")
-    public void onResponse(Message<String> message) {
-        String type = String.valueOf(message.getHeaders().get(HDR_TYPE));
-        if (!TYPE_RESULT.equals(type)) return;
+    public void onResponse(Message<CreateIdeaReportMetadataResponseDto> message) {
+        MessageHeaders headers = message.getHeaders();
+        String type = String.valueOf(headers.get(HDR_MESSAGE_TYPE));
+        String id = String.valueOf(headers.get("jobId"));
 
-        String inputIdStr = String.valueOf(message.getHeaders().get(HDR_INPUT_ID));
-        if (inputIdStr == null || "null".equals(inputIdStr)) return;
+        if (!MESSAGE_TYPE.equals(type) || id == null || id.equals("null")) {
+            log.debug("무시: type={}, jobId={}", type, id);
+            return;
+        }
 
-        UUID inputId = UUID.fromString(inputIdStr);
+        UUID jobId = UUID.fromString(id);
+        CreateIdeaReportMetadataResponseDto payload = message.getPayload();
+
         try {
-            AiIdeaReportResultMessage resultMessage = om.readValue(message.getPayload(), AiIdeaReportResultMessage.class);
-            AiIdeaReportResultMessage.ReportSummary summary = resultMessage.getSummaryReport().getReportSummary();
-            AiIdeaReportResultMessage.EvaluationScoresDto scores = summary.getEvaluationScores();
-
-            IdeaReportInput ideaReportInput = inputRepository.findById(inputId).orElse(null);
-            if (ideaReportInput == null) {
-                log.warn("응답 수신했지만 inputId를 못찾음: {}", inputId);
+            IdeaReportTask task = ideaReportTaskRepository.findById(jobId).orElse(null);
+            if (task == null) {
+                log.warn("응답 수신했지만 jobId를 못찾음: {}", jobId);
                 return;
             }
 
-            IdeaReportResult result = IdeaReportResult.builder()
-                    .evaluationScores(
-                            new EvaluationScores(
-                                    scores.getSimilarity(),
-                                    scores.getCreativity(),
-                                    scores.getFeasibility()
-                            )
-                    )
-                    .totalSimilarCases(summary.getTotalSimilarCases())
-                    .analysisNarrative(summary.getAnalysisNarrative())
-                    .build();
+            task.setStatus(IdeaReportTask.Status.SUCCEEDED);
+            task.setResultJson(objectMapper.writeValueAsString(payload));
+            ideaReportTaskRepository.save(task);
 
-            resultMessage.getDetailedReport().getDetailedResults().forEach(resultItem -> {
-                IdeaReportResultItem item = IdeaReportResultItem.builder()
-                        .result(result)
-                        .sourceType(resultItem.getSourceType() == null ? "-" : resultItem.getSourceType())
-                        .title(resultItem.getTitle() == null ? "-" : resultItem.getTitle())
-                        .link(resultItem.getLink() == null ? "-" : resultItem.getLink())
-                        .thumbnail(resultItem.getThumbnail() == null ? "-" : resultItem.getThumbnail())
-                        .summary(resultItem.getSummary() == null ? "-" : resultItem.getSummary())
-                        .score(resultItem.getScore() == null ? "0" : String.valueOf(resultItem.getScore()))
-                        .insight(resultItem.getInsight() == null ? "-" : resultItem.getInsight())
-                        .build();
-                result.getIdeaReportResultItems().add(item);
-            });
-
-            ideaReportInput.setResult(result);
-            IdeaReportInput saved = inputRepository.save(ideaReportInput);
-
-            IdeaReportMetadataResponseDto responseDto =
-                    IdeaReportMetadataResponseDto.of(
-                            saved.getId(),
-                            scores.getSimilarity(),
-                            scores.getCreativity(),
-                            scores.getFeasibility(),
-                            summary.getAnalysisNarrative(),
-                            saved.getResult().getIdeaReportResultItems().stream()
-                                    .sorted(
-                                            Comparator.comparingDouble(
-                                                    (IdeaReportResultItem e) -> Double.parseDouble(e.getScore())
-                                            ).reversed()
-                                    )
-                                    .map(item -> IdeaReportMetadataResponseDto.ResultItem.builder()
-                                            .id(item.getId())
-                                            .sourceType(item.getSourceType())
-                                            .title(item.getTitle())
-                                            .link(item.getLink())
-                                            .thumbnail(item.getThumbnail())
-                                            .summary(item.getSummary())
-                                            .score(item.getScore())
-                                            .insight(item.getInsight())
-                                            .build()
-                                    ).toList()
-                    );
-
-            messagingTemplate.convertAndSend("/topic/idea-report/metadata/" + inputId, responseDto);
-            log.info("웹소켓 푸시 완료: /topic/idea-report/metadata/{}", inputId);
+            messagingTemplate.convertAndSend("/topic/idea-reports/metadata/" + jobId, payload);
+            log.info("웹소켓 푸시 완료: /topic/idea-reports/metadata/{}", jobId);
         } catch (Exception e) {
-            log.error("AI 응답 처리 실패", e);
+            log.error("응답 처리 중 오류", e);
             throw new RuntimeException(e);
         }
     }
