+++ src/main/java/com/github/ideantifyserver/domain/ideareport/sqs/IdeaReportMetadataSqsGateway.java
@@ -1,7 +1,8 @@
 package com.github.ideantifyserver.domain.ideareport.sqs;
 
-import com.github.ideantifyserver.domain.ideareport.dto.request.CreateIdeaReportMetadataRequestDto;
 import io.awspring.cloud.sqs.operations.SqsTemplate;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
@@ -11,22 +12,30 @@ import java.util.UUID;
 @Component
 @RequiredArgsConstructor
 public class IdeaReportMetadataSqsGateway {
-
     private final SqsTemplate sqsTemplate;
 
     @Value("${app.sqs.idea-report-metadata.request-queue}")
     private String requestQueue;
 
-    private static final String HDR_TYPE = "messageType";
-    private static final String HDR_INPUT_ID = "inputId";
+    private static final String HDR_MESSAGE_TYPE = "messageType";
+    private static final String MESSAGE_TYPE     = "IDEA_REPORT";
 
-    public void publish(UUID inputId, CreateIdeaReportMetadataRequestDto dto, String responseQueue) {
-        sqsTemplate.send(to -> to
+    public void requestMetadata(UUID jobId, String query, String responseQueue) {
+        IdeaReportMetadataRequestMessage payload = new IdeaReportMetadataRequestMessage(jobId, query, responseQueue);
+
+        sqsTemplate.send(t -> t
                 .queue(requestQueue)
-                .payload(dto)
-                .header(HDR_TYPE, "IDEA_REPORT_REQUEST")
-                .header(HDR_INPUT_ID, inputId.toString())
-                .header("responseQueue", responseQueue)
+                .payload(payload)
+                .header(HDR_MESSAGE_TYPE, MESSAGE_TYPE)
+                .header("jobId", jobId.toString())
         );
+    }
+
+    @Getter
+    @AllArgsConstructor
+    public static class IdeaReportMetadataRequestMessage {
+        private UUID jobId;
+        private String query;
+        private String replyTo;
     }
 }
