+++ src/main/java/com/github/ideantifyserver/domain/ideareport/sqs/IdeaReportSqsGateway.java
@@ -1,8 +1,7 @@
 package com.github.ideantifyserver.domain.ideareport.sqs;
 
+import com.github.ideantifyserver.domain.ideareport.dto.request.CreateIdeaReportRequestDto;
 import io.awspring.cloud.sqs.operations.SqsTemplate;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
@@ -12,30 +11,22 @@ import java.util.UUID;
 @Component
 @RequiredArgsConstructor
 public class IdeaReportSqsGateway {
+
     private final SqsTemplate sqsTemplate;
 
     @Value("${app.sqs.idea-report.request-queue}")
     private String requestQueue;
 
-    private static final String HDR_MESSAGE_TYPE = "messageType";
-    private static final String MESSAGE_TYPE     = "IDEA_REPORT";
+    private static final String HDR_TYPE = "messageType";
+    private static final String HDR_INPUT_ID = "inputId";
 
-    public void requestReport(UUID jobId, String query, String responseQueue) {
-        IdeaReportRequestMessage payload = new IdeaReportRequestMessage(jobId, query, responseQueue);
-
-        sqsTemplate.send(t -> t
+    public void publish(UUID inputId, CreateIdeaReportRequestDto dto, String responseQueue) {
+        sqsTemplate.send(to -> to
                 .queue(requestQueue)
-                .payload(payload)
-                .header(HDR_MESSAGE_TYPE, MESSAGE_TYPE)
-                .header("jobId", jobId.toString())
+                .payload(dto)
+                .header(HDR_TYPE, "IDEA_REPORT_REQUEST")
+                .header(HDR_INPUT_ID, inputId.toString())
+                .header("responseQueue", responseQueue)
         );
-    }
-
-    @Getter
-    @AllArgsConstructor
-    public static class IdeaReportRequestMessage {
-        private UUID jobId;
-        private String query;
-        private String replyTo;
     }
 }
